back/requirements.txt
Django>=5.0.6
psycopg2-binary==2.9.9
djangorestframework==3.15.2
Pillow